scsi: Don't fill dirty page content in the SG_IO buffer
BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1168787

In sg_io, blk_rq_map_user{,_iov} may allocate a set of bounce buffer
pages to do the bio, if it finds the user buffer cannot be directly
mapped. But the allocated pages are not cleared.

If the bounce buffer is also not written to by device, garbage data is
left, and copied back to user in blk_rq_unmap_user.

Fix this by clearing the allocated pages. This also eliminates the risk
of leaking sensitive information to userspace, which may have a security
impact.

To reproduce with a virtio-scsi device, run this in guest:

  # sg_raw /dev/sda -r 54 0x12  0 1 0 36

This is not specific to virtualized devices, the conditions for the
leak are:

  1) The command's RW type is read.

  2) The device doesn't write any data to the provided buffer.

Signed-off-by: Fam Zheng <famz@redhat.com>