sd: Return error in sd_revalidate_disk if read capacity failed
When the read capacity commands failed, we should return an error to
caller instead of silently continuing as normal.

Most importantly, this fixes the error code of BLKRRPART ioctl. Also if
the device is down, the following commands also likely to time
out, and we could waste more time than necessary.

Signed-off-by: Fam Zheng <famz@redhat.com>